services/feature-extractor/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
tensorflow==2.15.0
numpy==1.24.3
Pillow==10.1.0
kafka-python==2.0.2
minio==7.2.0
pydantic==2.5.0
